requirements.txt
pip
setuptools
wheel
virtualenv
reportlab
pillow
inotify-simple
pycups
PyPDF2
requests
beautifulsoup4
flask
flask-sock
pytz
google-auth-oauthlib
google-auth
google-auth-httplib2
google-api-python-client
python-dateutil
pyyaml
boto3
